refactor: 拆分线程动作大文件并整理回归测试
本次提交聚焦 useThreadActions 相关历史大文件治理，在不改变现有线程动作对外 contract 的前提下，对纯 helper 与回归测试做低风险拆分。\n\n主要改动包括：\n- 将 useThreadActions.ts 中与线程筛选、rewind 目标解析、Gemini session 归一化、collab 关系恢复、队列替换判断等纯函数逻辑抽离到新的 useThreadActions.helpers.ts，减少主 hook 文件体量并降低后续维护时的上下文耦合。\n- 保持 useThreadActions 主文件中的 orchestration、state/ref 管理、runtime 调用与 dispatch 行为不变，仅调整 helper import 与调用位置，避免引入跨层 contract 漂移。\n- 将 Claude/Codex rewind 相关回归测试从 useThreadActions.test.tsx 拆分到新的 useThreadActions.rewind.test.tsx，保留原断言语义，覆盖 Claude session fork、从 message rewind、workspace 文件恢复、失败回滚与 files-only/messages-only 模式等场景。\n- 同步清理测试文件中拆分后未再使用的 import，并修复 helper 外移过程中遗漏的 sizeBytes 归一化调用，确保 listThreadsForWorkspace 的 Claude session 合并路径保持正确。\n\n涉及模块：\n- src/features/threads/hooks/useThreadActions.ts\n- src/features/threads/hooks/useThreadActions.helpers.ts\n- src/features/threads/hooks/useThreadActions.test.tsx\n- src/features/threads/hooks/useThreadActions.rewind.test.tsx\n\n主要影响：\n- useThreadActions.ts 从 3088 行降至 2447 行。\n- useThreadActions.test.tsx 从 3283 行降至 2336 行。\n- large-file hard gate 中这两个历史超限文件已回到阈值内。\n- 线程动作、rewind 与列表恢复行为保持不变，后续 review 与迭代的定位成本更低。\n\n验证结果：\n- 通过: npx vitest run src/features/threads/hooks/useThreadActions.test.tsx src/features/threads/hooks/useThreadActions.rewind.test.tsx\n- 通过: npm run typecheck\n- 通过: npm run check:large-files
